backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
sqlalchemy>=2.0.36,<3
pymysql>=1.1,<2
redis>=5.2,<6
python-dotenv>=1.0,<2
pydantic>=2.10,<3
pydantic-settings>=2.6,<3
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]>=1.7.4,<2
python-multipart>=0.0.20,<1
httpx>=0.28,<1
cryptography>=43,<46
